+++ tests/Factory/Environment/EnvironmentFactoryTest.php
@@ -4,8 +4,7 @@ declare(strict_types = 1);
 namespace Tests\Innmind\Http\Factory\Environment;
 
 use Innmind\Http\{
-    Factory\Environment\EnvironmentFactory,
-    Factory\EnvironmentFactory as EnvironmentFactoryInterface,
+    Factory\EnvironmentFactory,
     ServerRequest\Environment,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
@@ -15,8 +14,6 @@ class EnvironmentFactoryTest extends TestCase
     public function testMake()
     {
         $f = EnvironmentFactory::default();
-
-        $this->assertInstanceOf(EnvironmentFactoryInterface::class, $f);
 
         $e = ($f)();