+++ CHANGELOG.md
@@ -9,10 +9,12 @@
 - Requires `innmind/io:~3.0`
 - `Innmind\Http\Sender::__invoke()` now returns `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`
 - `Innmind\Http\Factory\CookiesFactory` is now a final class
+- `Innmind\Http\Factory\EnvironmentFactory` is now a final class
 
 ### Removed
 
 - `Innmind\Http\Factory\Cookies\CookiesFactory`
+- `Innmind\Http\Factory\Environment\EnvironmentFactory`
 
 ### Fixed